Workevent refactoring
Moved methods related to getting all the assigned jobs of a specific user to more suitable projects
